New document: Annotation Cheat Sheet
A couple people have asked whether we have a list of Spring annotations. We didn't, so I combed the existing documents to make one. I'm sure I missed a few, and I might be wrong about what a couple of them do, so this is a draft rather than a finished document.